oc_backend/requirements.txt
Django ~= 1.11
amqp ~= 2.2
celery ~= 4.1
djangorestframework ~= 3.7
docker ~= 3.1
psycopg2 ~= 2.7
django-cors-headers ~= 2.1
redis ~= 2.10